models/requirements.txt
python-dotenv
ultralytics
flask
requests